ファイルが存在せず、履歴もない場合に無限ループに至っていた不具合を修正、ファイルダウンロードの不具合の修正

+++ src/Docms.Client/FileSyncing/FileSynchronizer.cs
@@ -85,7 +85,7 @@ namespace Docms.Client.FileSyncing
                         }
                     }
                 }
-                else
+                else if (file.Path != null && file.Path == path)
                 {
                     await _client.DeleteDocumentAsync(path);
                 }
